feat: SJRA-633 split integration test in multiple files
